requirements.txt
tensorflow==2.15.0
torch==2.3.0
transformers==4.41.2
opencv-python
numpy==1.24.4
pyyaml==6.0.2
scikit-learn==1.3.2
matplotlib==3.8.0
pandas==2.1.3
seaborn==0.12.2